fix compilation error on newer gcc versions (>= 4.6.1)

+++ src/gui-utils.h
@@ -213,7 +213,7 @@ class combo_box : public widget {
 		SDL_Surface* surf;
 		int expanded_index;
 		unsigned int chosen_index;
-		unsigned int max_items_expanded;
+		size_t max_items_expanded;
 		unsigned int num_items_expanded;
 };
